grpc-service/requirements.txt
grpcio
grpcio-tools
googleapis-common-protos
sqlalchemy
psycopg2